Create a test for createFighter function and organize the project
